Create README.md
This is was a group project between the iOS class and the Ruby Class  at the end of our 4th week to make a dictionary of all the characters in star wars using Wookiepedia.com. The app worked well, but my Ruby partner deleted our database of characters to make room on his server for other projects. There should be a 4 characters left to search for (Han Solo, Padme, Yoda, and Luke Skywalker).